+++ shells/powershell.go
@@ -83,7 +83,7 @@ func (b *PowerShell) GenerateScript(build *common.Build) (*common.ShellScript, e
 			continue
 		}
 
-		if !build.Runner.DisableVerbose {
+		if !helpers.BoolOrDefault(build.Runner.DisableVerbose, false) {
 			b.writeCommand(w, "echo \"%s\"", command)
 		}
 		b.writeCommandChecked(w, "%s", command)