+++ ssh/ssh_config.go
@@ -1,8 +1,8 @@
 package ssh
 
 type Config struct {
-	User     string `toml:"user" json:"user"`
-	Password string `toml:"password" json:"password"`
-	Host     string `toml:"host" json:"host"`
-	Port     string `toml:"port" json:"port"`
+	User     *string `toml:"user" json:"user"`
+	Password *string `toml:"password" json:"password"`
+	Host     *string `toml:"host" json:"host"`
+	Port     *string `toml:"port" json:"port"`
 }