+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v 0.3.0
 - Build path includes repository hostname
 - Support TLS connection with Docker
 - Default concurrent limit is set to number of CPUs
+- Make most of the config options optional
 
 v 0.2.0
 - Added delete and verify commands